+++ module/actor/actor-active-effects.mjs
@@ -1,6 +1,23 @@
 import { HEROSYS } from "../herosystem6e.mjs";
 import { RoundFavorPlayerUp } from "../utility/round.mjs";
 
+export class HeroSystem6eActorActiveEffectsSystemData extends foundry.abstract.TypeDataModel {
+    static defineSchema() {
+        const fields = foundry.data.fields;
+        return {
+            // Make sure active-effect-config.hbs has all these fields so they don't get lost during editing
+            changes: new fields.ArrayField(
+                new fields.SchemaField({
+                    seconds: new fields.NumberField({ integer: true }),
+                    activePoints: new fields.NumberField({ integer: false }),
+                    source: new fields.StringField(),
+                    startTime: new fields.NumberField({ integer: true }),
+                }),
+            ),
+        };
+    }
+}
+
 export class HeroSystem6eActorActiveEffects extends ActiveEffect {
     // static defineSchema() {
     //     const schema2 = this.schema; // foundry.deepClone(super.defineSchema());