+++ module/item/item-attack.mjs
@@ -2700,7 +2700,7 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
 
     // DRAIN
     const reductionChatMessages = [];
-    const reductionTargetActor = token.actor;
+    //const reductionTargetActor = token.actor;
     for (const reduce of reducesArray) {
         reductionChatMessages.push(
             await performAdjustment(
@@ -2710,7 +2710,7 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
                 defense,
                 damageDetail.effects,
                 false,
-                reductionTargetActor,
+                token,
                 action,
             ),
         );
@@ -2721,7 +2721,7 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
 
     // AID
     const enhancementChatMessages = [];
-    const enhancementTargetActor = adjustmentItem.system.XMLID === "TRANSFER" ? adjustmentItem.actor : token.actor;
+    //const enhancementTargetActor = adjustmentItem.system.XMLID === "TRANSFER" ? adjustmentItem.actor : token.actor;
     for (const enhance of enhancesArray) {
         const simplifiedHealing = adjustmentItem.system.XMLID === "HEALING" && enhance.match(/simplified/i);
 
@@ -2735,7 +2735,7 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
                     "None - Beneficial",
                     "",
                     false,
-                    enhancementTargetActor,
+                    token,
                 ),
             );
             // BODY
@@ -2747,7 +2747,7 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
                     "None - Beneficial",
                     "",
                     false,
-                    enhancementTargetActor,
+                    token,
                 ),
             );
             adjustmentItemTags.push({
@@ -2767,7 +2767,8 @@ async function _onApplyAdjustmentToSpecificToken(adjustmentItem, token, damageDe
                     "None - Beneficial",
                     "",
                     false,
-                    enhancementTargetActor,
+                    token,
+                    action,
                 ),
             );
             adjustmentItemTags.push({ name: enhance });